Extract newsletter code from ingestion processor for separation of concerns
