+++ services/newsletter_processor/requirements.txt
@@ -1,0 +1,3 @@
+boto3>=1.34.0
+python-dotenv>=1.0.0
+python-docx>=1.1.0
